+++ demo-one/src/main/java/com/example/customer/CustomerDao.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 
 public interface CustomerDao {
     List<Customer> selectAllCustomers();
-    Optional<Customer> selectCustomerById(Integer id);
+    Optional<Customer> selectCustomerById(Long id);
     Customer insertCustomer(Customer customer);
     boolean existsCustomerWithEmail(String email);
-    void deleteCustomerById(Integer id);
+    void deleteCustomerById(Long id);
     void updateCustomer(Customer customer);
 ;}
