+++ demo-one/src/main/java/com/example/customer/CustomerRepository.java
@@ -2,8 +2,7 @@ package com.example.customer;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
     boolean existsCustomerByEmail(String email);
-    void deleteById(Integer id);
-
+    void deleteById(Long id);
 }
